14/1/24

+++ src/main/java/com/example/project_client/view/controller/Truong/ingredientView.java
@@ -7,13 +7,13 @@ import com.example.project_client.router.Router;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import lombok.Getter;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 
 public final class ingredientView {
@@ -47,6 +47,19 @@ public final class ingredientView {
         Router.switchTo(Pages.MAIN_VIEW);
     }
     @FXML
+    private void readIngredient() {
+        try {
+            ingredient = tableView.getSelectionModel().getSelectedItem();
+            if(ingredient == null) {
+                throw new Exception("Please choose Ingredient to view information");
+            }
+            Router.switchTo(Pages.READ_INGREDIENT);
+        }
+        catch (Exception e) {
+            raiseAlert(e.getMessage());
+        }
+    }
+    @FXML
     public void addIngredient() throws IOException {
         Router.switchTo(Pages.ADD_INGREDIENT);
     }
@@ -54,29 +67,68 @@ public final class ingredientView {
     public void changeIngredient() {
         try {
             ingredient = tableView.getSelectionModel().getSelectedItem();
+            if(ingredient == null) {
+                throw new Exception("Please choose Ingredient to change");
+            }
             Router.switchTo(Pages.CHANGE_INGREDIENT);
         }
         catch (Exception e) {
-            System.out.println("ERROR");
+            raiseAlert(e.getMessage());
         }
     }
     @FXML
     public void deleteIngredient() {
         try {
             ingredient = tableView.getSelectionModel().getSelectedItem();
-            IngredientRepository.deleteIngredient(ingredient.getId().toString());
-            tableList.remove(ingredient);
+            if(ingredient == null) {
+                throw new Exception("Please choose Ingredient to delete");
+            }
+            askDelete();
         }
         catch (Exception e){
-            System.out.println("Please choose Ingredient to delete");
+            raiseAlert(e.getMessage());
         }
+    }
+    private void askDelete() throws Exception {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation");
+        alert.setHeaderText("Do you want to delete Ingredient");
+        alert.setContentText("choose your option");
+
+        ButtonType buttonTypeYes = new ButtonType("Yes", ButtonBar.ButtonData.YES);
+        ButtonType buttonTypeNo = new ButtonType("No", ButtonBar.ButtonData.NO);
+
+        alert.getButtonTypes().setAll(buttonTypeYes, buttonTypeNo);
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        String message;
+        if(result.get().getButtonData() == ButtonBar.ButtonData.YES){
+            try {
+                IngredientRepository.deleteIngredient(ingredient.getId().toString());
+                tableList.remove(ingredient);
+                message = "Deleted";
+            }
+            catch (Exception e) {
+                message = "Cannot delete";
+            }
+        } else {
+            message = "Cancel delete";
+        }
+    }
+    private void raiseAlert(String alertText){
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle("Information");
+        alert.setHeaderText("Notification");
+        alert.setContentText(alertText);
+        alert.show();
     }
     private void setTableView() throws IOException {
         ingredients = IngredientRepository.getIngredientsApi();
         tableList = FXCollections.observableArrayList(ingredients);
         tableView.setItems(tableList);
     }
-    private void setColumn() throws IOException {
+    private void setColumn() {
         id.setCellValueFactory(new PropertyValueFactory<>("id"));
         name.setCellValueFactory(new PropertyValueFactory<>("name"));
         unit.setCellValueFactory(new PropertyValueFactory<>("unit"));
